youtube_shorts/requirements.txt
google-adk
python-dotenv
google-generativeai
openai
